+++ mutator/mutator.go
@@ -11,6 +11,7 @@ func Mutate(p *pkg.Package) *pkg.Package {
 		combineErrorsWithDefault,
 		inlinePrimitiveTypes,
 		inlineResponseStructs,
+		hoistEmbeddedStuctFields,
 		removeUnusedDecls,
 	}
 
@@ -62,7 +63,7 @@ func inlinePrimitiveTypes(p *pkg.Package) *pkg.Package {
 }
 
 func replaceType(p *pkg.Package, old, new pkg.Type) *pkg.Package {
-	return walkTypes(p, func(t pkg.Type) pkg.Type {
+	return walkTypes(p, func(t pkg.Type, _ typeContext) pkg.Type {
 		if old.Equal(t) {
 			return new
 		}
@@ -97,8 +98,12 @@ func inlineResponseStructs(p *pkg.Package) *pkg.Package {
 	for _, d := range p.TypeDecls {
 		di := pkg.IdentType{Name: d.Name}
 		if pc, ok := ctxs[di]; ok {
-			d.Type = recurseType(d.Type, func(t pkg.Type) pkg.Type {
+			d.Type = recurseType(d.Type, decl, func(t pkg.Type, c typeContext) pkg.Type {
 				if t == d.Type {
+					return t
+				}
+
+				if c&embeddedStruct > 0 {
 					return t
 				}
 
@@ -109,6 +114,78 @@ func inlineResponseStructs(p *pkg.Package) *pkg.Package {
 				}
 
 				return t
+			})
+		}
+	}
+
+	return p
+}
+
+// hoistEmbeddedStuctFields finds any embedded structs from allOf schemas, and
+// hoists their fields into the embedding struct, when the embedded struct is not
+// referenced elsewhere.
+func hoistEmbeddedStuctFields(p *pkg.Package) *pkg.Package {
+	// XXX dedupe this code with inlineResponseStructs
+	ctxs := make(map[pkg.IdentType]struct {
+		c typeContext
+		n int
+	}, len(p.TypeDecls))
+
+	reachDecls(p, func(c typeContext, i *pkg.IdentType) bool {
+		v := ctxs[*i]
+		v.c |= c
+		v.n++
+		ctxs[*i] = v
+		return true
+	})
+
+	for _, d := range p.TypeDecls {
+		di := pkg.IdentType{Name: d.Name}
+		if pc, ok := ctxs[di]; ok {
+			d.Type = recurseType(d.Type, decl, func(t pkg.Type, c typeContext) pkg.Type {
+				st, ok := t.(*pkg.StructType)
+				if !ok {
+					return t
+				}
+
+				var nf []pkg.Field
+
+				for i := range st.Fields {
+					f := st.Fields[i]
+					if f.ID != "" {
+						nf = append(nf, f)
+						continue
+					}
+
+					var embedded *pkg.StructType
+					switch ft := f.Type.(type) {
+					case *pkg.IdentType:
+						if cc, ok := ctxs[*ft]; !ok || pc.n < cc.n {
+							nf = append(nf, f)
+							continue
+						}
+
+						resolved := resolve(p, ft)
+						if rst, ok := resolved.(*pkg.StructType); ok {
+							embedded = rst
+						} else {
+							nf = append(nf, f)
+							continue
+						}
+					case *pkg.StructType:
+						embedded = ft
+					default:
+						nf = append(nf, f)
+						continue
+					}
+
+					for j := range embedded.Fields {
+						nf = append(nf, embedded.Fields[j])
+					}
+				}
+
+				st.Fields = nf
+				return st
 			})
 		}
 	}
@@ -157,7 +234,7 @@ func removeUnusedDecls(p *pkg.Package) *pkg.Package {
 	return p
 }
 
-type typeContext byte
+type typeContext uint16
 
 const (
 	decl typeContext = 1 << iota
@@ -166,8 +243,11 @@ const (
 	methodReturn
 	methodError
 
+	structField
+	embeddedStruct
+
 	none typeContext = 0
-	any  typeContext = 0xFF
+	any  typeContext = 0xFFFF
 )
 
 type stackItem struct {
@@ -226,30 +306,30 @@ func reachDecls(p *pkg.Package, fn func(typeContext, *pkg.IdentType) bool) {
 	}
 }
 
-func walkTypes(p *pkg.Package, fn func(pkg.Type) pkg.Type) *pkg.Package {
+func walkTypes(p *pkg.Package, fn func(pkg.Type, typeContext) pkg.Type) *pkg.Package {
 	for i, d := range p.TypeDecls {
-		d.Type = recurseType(d.Type, fn)
+		d.Type = recurseType(d.Type, decl, fn)
 		p.TypeDecls[i] = d
 	}
 
-	for i, iter := range p.Iters {
-		iter.Return = recurseType(iter.Return, fn)
-		p.Iters[i] = iter
+	for i, itr := range p.Iters {
+		itr.Return = recurseType(itr.Return, iter, fn)
+		p.Iters[i] = itr
 	}
 
 	for _, c := range p.Clients {
 		for _, m := range c.Methods {
 			for i, param := range m.Params {
-				param.Type = recurseType(param.Type, fn)
+				param.Type = recurseType(param.Type, methodParam, fn)
 				m.Params[i] = param
 			}
 
 			for i, ret := range m.Return {
-				m.Return[i] = recurseType(ret, fn)
+				m.Return[i] = recurseType(ret, methodReturn, fn)
 			}
 
 			for k, e := range m.Errors {
-				m.Errors[k] = recurseType(e, fn)
+				m.Errors[k] = recurseType(e, methodError, fn)
 			}
 		}
 	}
@@ -257,25 +337,30 @@ func walkTypes(p *pkg.Package, fn func(pkg.Type) pkg.Type) *pkg.Package {
 	return p
 }
 
-func recurseType(typ pkg.Type, fn func(pkg.Type) pkg.Type) pkg.Type {
+func recurseType(typ pkg.Type, parentCtx typeContext, fn func(pkg.Type, typeContext) pkg.Type) pkg.Type {
 	switch t := typ.(type) {
 	case *pkg.StructType:
 		for i := range t.Fields {
-			t.Fields[i].Type = recurseType(t.Fields[i].Type, fn)
+			c := structField
+			if t.Fields[i].ID == "" {
+				c = embeddedStruct
+			}
+
+			t.Fields[i].Type = recurseType(t.Fields[i].Type, c, fn)
 		}
 	case *pkg.SliceType:
-		t.Type = recurseType(t.Type, fn)
+		t.Type = recurseType(t.Type, parentCtx, fn)
 	case *pkg.IterType:
-		t.Type = recurseType(t.Type, fn)
+		t.Type = recurseType(t.Type, parentCtx|iter, fn)
 	case *pkg.PointerType:
-		t.Type = recurseType(t.Type, fn)
+		t.Type = recurseType(t.Type, parentCtx, fn)
 	}
 
-	return fn(typ)
+	return fn(typ, parentCtx)
 }
 
 func eachIdent(typ pkg.Type, fn func(*pkg.IdentType)) {
-	recurseType(typ, func(t pkg.Type) pkg.Type {
+	recurseType(typ, any, func(t pkg.Type, _ typeContext) pkg.Type {
 		if i, ok := t.(*pkg.IdentType); ok {
 			fn(i)
 		}