+++ mutator/mutator_test.go
@@ -172,10 +172,10 @@ func TestInlineResponseStruct(t *testing.T) {
 		return pkg.Package{
 			TypeDecls: []pkg.TypeDecl{
 				{Name: "FieldThing", Type: &pkg.StructType{
-					Fields: []pkg.Field{{Type: &pkg.IdentType{Name: "string"}}},
+					Fields: []pkg.Field{{ID: "Field", Type: &pkg.IdentType{Name: "string"}}},
 				}},
 				{Name: "StructThing", Type: &pkg.StructType{
-					Fields: []pkg.Field{{Type: &pkg.IdentType{Name: "FieldThing"}}},
+					Fields: []pkg.Field{{ID: "Field", Type: &pkg.IdentType{Name: "FieldThing"}}},
 				}},
 			},
 			Clients: c,
@@ -186,12 +186,26 @@ func TestInlineResponseStruct(t *testing.T) {
 		return pkg.Package{
 			TypeDecls: []pkg.TypeDecl{
 				{Name: "FieldThing", Type: &pkg.StructType{
-					Fields: []pkg.Field{{Type: &pkg.IdentType{Name: "string"}}},
+					Fields: []pkg.Field{{ID: "Field", Type: &pkg.IdentType{Name: "string"}}},
 				}},
 				{Name: "StructThing", Type: &pkg.StructType{
-					Fields: []pkg.Field{{Type: &pkg.StructType{
-						Fields: []pkg.Field{{Type: &pkg.IdentType{Name: "string"}}},
+					Fields: []pkg.Field{{ID: "Field", Type: &pkg.StructType{
+						Fields: []pkg.Field{{ID: "Field", Type: &pkg.IdentType{Name: "string"}}},
 					}}},
+				}},
+			},
+			Clients: c,
+		}
+	}
+
+	embeddedStructPkg := func(c ...pkg.Client) pkg.Package {
+		return pkg.Package{
+			TypeDecls: []pkg.TypeDecl{
+				{Name: "FieldThing", Type: &pkg.StructType{
+					Fields: []pkg.Field{{ID: "Field", Type: &pkg.IdentType{Name: "string"}}},
+				}},
+				{Name: "StructThing", Type: &pkg.StructType{
+					Fields: []pkg.Field{{Type: &pkg.IdentType{Name: "FieldThing"}}},
 				}},
 			},
 			Clients: c,
@@ -202,13 +216,13 @@ func TestInlineResponseStruct(t *testing.T) {
 		return pkg.Package{
 			TypeDecls: []pkg.TypeDecl{
 				{Name: "FieldThing", Type: &pkg.StructType{
-					Fields: []pkg.Field{{Type: &pkg.IdentType{Name: "string"}}},
+					Fields: []pkg.Field{{ID: "Field", Type: &pkg.IdentType{Name: "string"}}},
 				}},
 				{Name: "StructThing", Type: &pkg.StructType{
-					Fields: []pkg.Field{{Type: &pkg.IdentType{Name: "FieldThing"}}},
+					Fields: []pkg.Field{{ID: "Field", Type: &pkg.IdentType{Name: "FieldThing"}}},
 				}},
 				{Name: "OtherStructThing", Type: &pkg.StructType{
-					Fields: []pkg.Field{{Type: &pkg.IdentType{Name: "FieldThing"}}},
+					Fields: []pkg.Field{{ID: "Field", Type: &pkg.IdentType{Name: "FieldThing"}}},
 				}},
 			},
 			Clients: c,
@@ -219,15 +233,15 @@ func TestInlineResponseStruct(t *testing.T) {
 		return pkg.Package{
 			TypeDecls: []pkg.TypeDecl{
 				{Name: "FieldThing", Type: &pkg.StructType{
-					Fields: []pkg.Field{{Type: &pkg.IdentType{Name: "string"}}},
+					Fields: []pkg.Field{{ID: "Field", Type: &pkg.IdentType{Name: "string"}}},
 				}},
 				{Name: "StructThing", Type: &pkg.StructType{
-					Fields: []pkg.Field{{Type: &pkg.StructType{
-						Fields: []pkg.Field{{Type: &pkg.IdentType{Name: "string"}}},
+					Fields: []pkg.Field{{ID: "Field", Type: &pkg.StructType{
+						Fields: []pkg.Field{{ID: "Field", Type: &pkg.IdentType{Name: "string"}}},
 					}}},
 				}},
 				{Name: "OtherStructThing", Type: &pkg.StructType{
-					Fields: []pkg.Field{{Type: &pkg.IdentType{Name: "FieldThing"}}},
+					Fields: []pkg.Field{{ID: "Field", Type: &pkg.IdentType{Name: "FieldThing"}}},
 				}},
 			},
 			Clients: c,
@@ -247,6 +261,15 @@ func TestInlineResponseStruct(t *testing.T) {
 			}}}),
 			basePkg(pkg.Client{Methods: []pkg.Method{{
 				Params: []pkg.Param{{Type: &pkg.IdentType{Name: "StructThing"}}},
+			}}}),
+		},
+
+		{"not inlined if embedded struct",
+			embeddedStructPkg(pkg.Client{Methods: []pkg.Method{{
+				Return: []pkg.Type{&pkg.IdentType{Name: "StructThing"}},
+			}}}),
+			embeddedStructPkg(pkg.Client{Methods: []pkg.Method{{
+				Return: []pkg.Type{&pkg.IdentType{Name: "StructThing"}},
 			}}}),
 		},
 
@@ -311,6 +334,176 @@ func TestInlineResponseStruct(t *testing.T) {
 	for _, tc := range tcs {
 		t.Run(tc.name, func(t *testing.T) {
 			out := inlineResponseStructs(&tc.in)
+			if !reflect.DeepEqual(*out, tc.out) {
+				t.Error("got:", out, "expected:", tc.out)
+			}
+		})
+	}
+}
+
+func TestHoistEmbeddedStructFields(t *testing.T) {
+	basePkg := func(c ...pkg.Client) pkg.Package {
+		return pkg.Package{
+			TypeDecls: []pkg.TypeDecl{
+				{Name: "FieldThing", Type: &pkg.StructType{
+					Fields: []pkg.Field{{ID: "Field", Type: &pkg.IdentType{Name: "string"}}},
+				}},
+				{Name: "StructThing", Type: &pkg.StructType{
+					Fields: []pkg.Field{{Type: &pkg.IdentType{Name: "FieldThing"}}},
+				}},
+			},
+			Clients: c,
+		}
+	}
+
+	namedFieldPkg := func(c ...pkg.Client) pkg.Package {
+		return pkg.Package{
+			TypeDecls: []pkg.TypeDecl{
+				{Name: "FieldThing", Type: &pkg.StructType{
+					Fields: []pkg.Field{{ID: "Field", Type: &pkg.IdentType{Name: "string"}}},
+				}},
+				{Name: "StructThing", Type: &pkg.StructType{
+					Fields: []pkg.Field{{ID: "Field", Type: &pkg.IdentType{Name: "FieldThing"}}},
+				}},
+			},
+			Clients: c,
+		}
+	}
+
+	hoistedPkg := func(c ...pkg.Client) pkg.Package {
+		return pkg.Package{
+			TypeDecls: []pkg.TypeDecl{
+				{Name: "FieldThing", Type: &pkg.StructType{
+					Fields: []pkg.Field{{ID: "Field", Type: &pkg.IdentType{Name: "string"}}},
+				}},
+				{Name: "StructThing", Type: &pkg.StructType{
+					Fields: []pkg.Field{{ID: "Field", Type: &pkg.IdentType{Name: "string"}}},
+				}},
+			},
+			Clients: c,
+		}
+	}
+
+	doublePkg := func(c ...pkg.Client) pkg.Package {
+		return pkg.Package{
+			TypeDecls: []pkg.TypeDecl{
+				{Name: "FieldThing", Type: &pkg.StructType{
+					Fields: []pkg.Field{{ID: "Field", Type: &pkg.IdentType{Name: "string"}}},
+				}},
+				{Name: "StructThing", Type: &pkg.StructType{
+					Fields: []pkg.Field{{Type: &pkg.IdentType{Name: "FieldThing"}}},
+				}},
+				{Name: "OtherStructThing", Type: &pkg.StructType{
+					Fields: []pkg.Field{{Type: &pkg.IdentType{Name: "FieldThing"}}},
+				}},
+			},
+			Clients: c,
+		}
+	}
+
+	hoistedDoublePkg := func(c ...pkg.Client) pkg.Package {
+		return pkg.Package{
+			TypeDecls: []pkg.TypeDecl{
+				{Name: "FieldThing", Type: &pkg.StructType{
+					Fields: []pkg.Field{{ID: "Field", Type: &pkg.IdentType{Name: "string"}}},
+				}},
+				{Name: "StructThing", Type: &pkg.StructType{
+					Fields: []pkg.Field{{ID: "Field", Type: &pkg.IdentType{Name: "string"}}},
+				}},
+				{Name: "OtherStructThing", Type: &pkg.StructType{
+					Fields: []pkg.Field{{Type: &pkg.IdentType{Name: "FieldThing"}}},
+				}},
+			},
+			Clients: c,
+		}
+	}
+
+	tcs := []struct {
+		name string
+		in   pkg.Package
+		out  pkg.Package
+	}{
+		{"empty", pkg.Package{}, pkg.Package{}},
+
+		{"not hoisted if not embedded struct",
+			namedFieldPkg(pkg.Client{Methods: []pkg.Method{{
+				Return: []pkg.Type{&pkg.IdentType{Name: "StructThing"}},
+			}}}),
+			namedFieldPkg(pkg.Client{Methods: []pkg.Method{{
+				Return: []pkg.Type{&pkg.IdentType{Name: "StructThing"}},
+			}}}),
+		},
+
+		{"hoisted in params",
+			basePkg(pkg.Client{Methods: []pkg.Method{{
+				Params: []pkg.Param{{Type: &pkg.IdentType{Name: "StructThing"}}},
+			}}}),
+			hoistedPkg(pkg.Client{Methods: []pkg.Method{{
+				Params: []pkg.Param{{Type: &pkg.IdentType{Name: "StructThing"}}},
+			}}}),
+		},
+
+		{"hoisted in return",
+			basePkg(pkg.Client{Methods: []pkg.Method{{
+				Return: []pkg.Type{&pkg.IdentType{Name: "StructThing"}},
+			}}}),
+			hoistedPkg(pkg.Client{Methods: []pkg.Method{{
+				Return: []pkg.Type{&pkg.IdentType{Name: "StructThing"}},
+			}}}),
+		},
+
+		{"hoisted in error",
+			basePkg(pkg.Client{Methods: []pkg.Method{{
+				Errors: map[int]pkg.Type{-1: &pkg.IdentType{Name: "StructThing"}},
+			}}}),
+			hoistedPkg(pkg.Client{Methods: []pkg.Method{{
+				Errors: map[int]pkg.Type{-1: &pkg.IdentType{Name: "StructThing"}},
+			}}}),
+		},
+
+		{"hosited if used from many calls",
+			basePkg(
+				pkg.Client{Methods: []pkg.Method{{
+					Return: []pkg.Type{&pkg.IdentType{Name: "StructThing"}},
+				}}},
+				pkg.Client{Methods: []pkg.Method{{
+					Return: []pkg.Type{&pkg.IdentType{Name: "StructThing"}},
+				}}},
+			),
+			hoistedPkg(
+				pkg.Client{Methods: []pkg.Method{{
+					Return: []pkg.Type{&pkg.IdentType{Name: "StructThing"}},
+				}}},
+				pkg.Client{Methods: []pkg.Method{{
+					Return: []pkg.Type{&pkg.IdentType{Name: "StructThing"}},
+				}}},
+			),
+		},
+
+		{"not hoisted if used in many structs",
+			doublePkg(pkg.Client{Methods: []pkg.Method{{Return: []pkg.Type{
+				&pkg.IdentType{Name: "StructThing"},
+				&pkg.IdentType{Name: "OtherStructThing"},
+			}}}}),
+			doublePkg(pkg.Client{Methods: []pkg.Method{{Return: []pkg.Type{
+				&pkg.IdentType{Name: "StructThing"},
+				&pkg.IdentType{Name: "OtherStructThing"},
+			}}}}),
+		},
+
+		{"hoisted if used in many structs but only one reachable",
+			doublePkg(pkg.Client{Methods: []pkg.Method{{
+				Return: []pkg.Type{&pkg.IdentType{Name: "StructThing"}},
+			}}}),
+			hoistedDoublePkg(pkg.Client{Methods: []pkg.Method{{
+				Return: []pkg.Type{&pkg.IdentType{Name: "StructThing"}},
+			}}}),
+		},
+	}
+
+	for _, tc := range tcs {
+		t.Run(tc.name, func(t *testing.T) {
+			out := hoistEmbeddedStuctFields(&tc.in)
 			if !reflect.DeepEqual(*out, tc.out) {
 				t.Error("got:", out, "expected:", tc.out)
 			}